+++ platformio/package/pack.py
@@ -22,7 +22,11 @@ import tempfile
 from platformio import fs
 from platformio.compat import IS_WINDOWS
 from platformio.package.exception import PackageException, UserSideException
-from platformio.package.manifest.parser import ManifestFileType, ManifestParserFactory
+from platformio.package.manifest.parser import (
+    LibraryPropertiesManifestParser,
+    ManifestFileType,
+    ManifestParserFactory,
+)
 from platformio.package.manifest.schema import ManifestSchema
 from platformio.package.meta import PackageItem
 from platformio.package.unpack import FileUnpacker
@@ -43,6 +47,7 @@ class PackagePacker(object):
         ".cache",
         "**/.cache",
         "**/__pycache__",
+        "**/*.pyc",
         # VCS
         ".git/",
         ".hg/",
@@ -50,16 +55,26 @@ class PackagePacker(object):
     ]
     EXCLUDE_EXTRA = [
         # Tests
-        "tests?",
+        "test",
+        "tests",
         # Docs
         "doc",
         "docs",
         "mkdocs",
+        "doxygen",
+        "*.doxyfile",
+        "html",
+        "media",
         "**/*.[pP][dD][fF]",
-        "**/*.[dD][oO][cC]?",
-        "**/*.[pP][pP][tT]?",
+        "**/*.[dD][oO][cC]",
+        "**/*.[dD][oO][cC][xX]",
+        "**/*.[pP][pP][tT]",
+        "**/*.[pP][pP][tT][xX]",
+        "**/*.[xX][lL][sS]",
+        "**/*.[xX][lL][sS][xX]",
         "**/*.[dD][oO][xX]",
-        "**/*.[hH][tT][mM]?",
+        "**/*.[hH][tT][mM]",
+        "**/*.[hH][tT][mM][lL]",
         "**/*.[tT][eE][xX]",
         "**/*.[jJ][sS]",
         "**/*.[cC][sS][sS]",
@@ -75,14 +90,13 @@ class PackagePacker(object):
         "**/*.[mM][pP][34]",
         "**/*.[pP][sS][dD]",
         "**/*.[wW][aA][wW]",
+        "**/*.sqlite",
     ]
     EXCLUDE_LIBRARY_EXTRA = [
         "assets",
         "extra",
+        "extras",
         "resources",
-        "html",
-        "media",
-        "doxygen",
         "**/build/",
         "**/*.flat",
         "**/*.[jJ][aA][rR]",
@@ -97,6 +111,7 @@ class PackagePacker(object):
     def __init__(self, package, manifest_uri=None):
         self.package = package
         self.manifest_uri = manifest_uri
+        self.manifest_parser = None
 
     @staticmethod
     def get_archive_name(name, version, system=None):
@@ -128,7 +143,8 @@ class PackagePacker(object):
                 src = tmp_dir
 
             src = self.find_source_root(src)
-            manifest = self.load_manifest(src)
+            self.manifest_parser = ManifestParserFactory.new_from_dir(src)
+            manifest = ManifestSchema().load_manifest(self.manifest_parser.as_dict())
             filename = self.get_archive_name(
                 manifest["name"],
                 manifest["version"],
@@ -143,11 +159,6 @@ class PackagePacker(object):
             return self._create_tarball(src, dst, manifest)
         finally:
             shutil.rmtree(tmp_dir)
-
-    @staticmethod
-    def load_manifest(src):
-        mp = ManifestParserFactory.new_from_dir(src)
-        return ManifestSchema().load_manifest(mp.as_dict())
 
     def find_source_root(self, src):
         if self.manifest_uri:
@@ -214,7 +225,9 @@ class PackagePacker(object):
         # exclude items declared in manifest
         result += ["-<%s>" % p for p in exclude or []]
         # apply extra excludes if no custom "export" field in manifest
-        if not include and not exclude:
+        if (not include and not exclude) or isinstance(
+            self.manifest_parser, LibraryPropertiesManifestParser
+        ):
             result += ["-<%s>" % p for p in exclude_extra]
         # automatically include manifests
         result += ["+<%s>" % p for p in self.INCLUDE_DEFAULT]