+++ platformio/project/helpers.py
@@ -14,12 +14,13 @@
 
 import json
 import os
+import subprocess
 from hashlib import sha1
 
 from click.testing import CliRunner
 
 from platformio import __version__, exception, fs
-from platformio.compat import IS_WINDOWS, hashlib_encode_data
+from platformio.compat import IS_MACOS, IS_WINDOWS, hashlib_encode_data
 from platformio.project.config import ProjectConfig
 
 
@@ -75,7 +76,15 @@ def get_default_projects_dir():
         ctypes.windll.shell32.SHGetFolderPathW(None, 5, None, 0, buf)
         docs_dir = buf.value
     except:  # pylint: disable=bare-except
-        pass
+        if not IS_MACOS:
+            try:
+                docs_dir = (
+                    subprocess.check_output(["xdg-user-dir", "DOCUMENTS"])
+                    .decode("utf-8")
+                    .strip()
+                )
+            except FileNotFoundError:  # command not found
+                pass
     return os.path.join(docs_dir, "PlatformIO", "Projects")